casting errors resolved

+++ HighwayPatrol.cpp
@@ -12,9 +12,9 @@ void HighwayPatrol::scanHighway(Highway* h)
 {
     std::cout << name << ": scanning highway for speeders" << std::endl;
 
-    for( int i = h->vehicles.size(); --i >= 0; )
+    for( int i = static_cast<int>(h->vehicles.size()); --i >= 0; )
     {
-        auto* v = h->vehicles[i];
+        auto* v = h->vehicles[static_cast<u_long>(i)];
         if( v->speed > h->speedLimit + 5 )
         {
             pullOver(v, v->speed > (h->speedLimit + 15), h );
